services/catalog-service/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.7.0
python-dotenv==1.0.1
httpx==0.27.0
PyJWT==2.8.0
